Added curl check
Added file check for curl

+++ cli-english.sh
@@ -1,7 +1,10 @@
 #! /bin/bash
 
 # Get word list file:
-curl -o /tmp/clienglishwordlist https://gnuadm.in/eng.md 2> /dev/null
+if [[ ! -f /tmp/clienglishwordlist ]]
+then
+	curl -o /tmp/clienglishwordlist https://gnuadm.in/eng.md 2> /dev/null
+fi
 
 # Get word count:
 wordcount=$(cat /tmp/clienglishwordlist | wc -l)
